pyproject.toml
[tool.poetry]
name = "flask-ml"
version = "0.1.0"
description = ""
authors = ["Eric Souza <ericsouza0801@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7"
scikit-learn = "^0.23.1"
flask = "^1.1.2"
python-dotenv = "^0.13.0"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
black = {version = "^19.10b0", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
